requirements.txt
pandas
numpy
scikit-learn
xgboost
google-api-python-client
beautifulsoup4
selenium
isodate
shap
matplotlib
seaborn
joblib
python-dotenv
webdriver-manager
